Organize core modules into package
